move code to server package
