+++ cmd/argo-watcher/server/argo_api.go
@@ -1,4 +1,4 @@
-package main
+package server
 
 import (
 	"bytes"
