+++ cmd/argo-watcher/server/argo_status_updater_test.go
@@ -1,4 +1,4 @@
-package main
+package server
 
 import (
 	"fmt"
