+++ cmd/argo-watcher/server/router_test.go
@@ -1,4 +1,4 @@
-package main
+package server
 
 import (
 	"fmt"
