+++ cmd/argo-watcher/server/server_test.go
@@ -1,4 +1,4 @@
-package main
+package server
 
 import (
 	"testing"
